Refactor the 'Sweeper' interface
- Remove the no-op implementations of Sweeper and SweepTaskRunner
- Make Sweeper an enum instead of an interface
- Make SweepTaskRunner a concrete class
- Change the timestamp supplier type from Supplier<Long> to LongSupplier

I'm working on a sweep performance project which requires a lot of things
to be changed. The current "polymorphic" design of Sweeper doesn't quite
work anymore.

[no release notes]
