--- atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/sweep/NoOpSweepTaskRunner.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2015 Palantir Technologies
- *
- * Licensed under the BSD-3 License (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://opensource.org/licenses/BSD-3-Clause
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.palantir.atlasdb.sweep;
-
-import com.palantir.atlasdb.keyvalue.api.SweepResults;
-import com.palantir.atlasdb.keyvalue.api.TableReference;
-import com.palantir.atlasdb.protos.generated.TableMetadataPersistence.SweepStrategy;
-
-public class NoOpSweepTaskRunner implements SweepTaskRunner {
-    @Override
-    public SweepResults dryRun(TableReference tableRef, int rowBatchSize, int cellBatchSize, byte[] startRow) {
-        return SweepResults.createEmptySweepResult(getSweepTimestamp(SweepStrategy.NOTHING));
-    }
-
-    @Override
-    public SweepResults run(TableReference tableRef, int rowBatchSize, int cellBatchSize, byte[] startRow) {
-        return SweepResults.createEmptySweepResult(getSweepTimestamp(SweepStrategy.NOTHING));
-    }
-
-    @Override
-    public long getSweepTimestamp(SweepStrategy sweepStrategy) {
-        return 0;
-    }
-}
